rename hyp3_gamma.hyp3_rtc_gamma to hyp3_gamma.rtc
